Rearranged mobile routes to be more manageable
